More function conversion
* Converted many functions in skills
* Added more mud_progs and expanded SpeechProg
* Added projectile values to ObjectTemplates

+++ Source/SmaugCS/fight.cs
@@ -788,7 +788,7 @@ namespace SmaugCS
         }
 
         public static ReturnTypes projectile_hit(CharacterInstance ch, CharacterInstance victim, ObjectInstance wield,
-            ObjectInstance projectile, short dist)
+            ObjectInstance projectile, int dist)
         {
             if (projectile == null)
                 return (int) ReturnTypes.None;
